Move all matrix stuff into jaxfun/la
