+++ src/jaxfun/la/__init__.py
@@ -1,3 +1,4 @@
+from .blocktpmatrix import BlockTPMatrix as BlockTPMatrix
 from .diamatrix import (
     DiaMatrix as DiaMatrix,
     diags as diags,
@@ -6,3 +7,12 @@ from .diamatrix import (
 from .matrix import Matrix as Matrix
 from .matrixprotocol import MatrixProtocol as MatrixProtocol
 from .pinned import PinnedSystem as PinnedSystem
+from .tensormatrix import TensorMatrix as TensorMatrix
+from .tpmatrix import (
+    TPMatrices as TPMatrices,
+    TPMatrix as TPMatrix,
+    tpmats_to_kron as tpmats_to_kron,
+    tpmats_to_scipy_kron as tpmats_to_scipy_kron,
+    tpmats_to_scipy_sparse as tpmats_to_scipy_sparse,
+    vec as vec,
+)
